requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
jinja2
elevenlabs
python-dotenv
Pillow
requests
